fix(initialization): move InitializationTask class to another file
